move existing popover code to gtk_since_3_1*
