+++ gtk/gtk_since_3_12.go
@@ -1,3 +1,19 @@
+// Copyright (c) 2013-2014 Conformal Systems <info@conformal.com>
+//
+// Wrapper for GtkPopover originated from: http://opensource.conformal.com/
+//
+// Permission to use, copy, modify, and distribute this software for any
+// purpose with or without fee is hereby granted, provided that the above
+// copyright notice and this permission notice appear in all copies.
+//
+// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+// WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+// MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+// ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+// WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+// ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+// OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
 // +build !gtk_3_6,!gtk_3_8,!gtk_3_10
 // not use this: go build -tags gtk_3_8'. Otherwise, if no build tags are used, GTK 3.10
 
@@ -23,11 +39,18 @@ func init() {
 		// Objects/Interfaces
 		{glib.Type(C.gtk_flow_box_get_type()), marshalFlowBox},
 		{glib.Type(C.gtk_flow_box_child_get_type()), marshalFlowBoxChild},
+		{glib.Type(C.gtk_popover_get_type()), marshalPopover},
 	}
 	glib.RegisterGValueMarshalers(tm)
 
-	WrapMap["GtkFlowBox"] = wrapFlowBox
-	WrapMap["GtkFlowBoxChild"] = wrapFlowBoxChild
+	//Contribute to casting
+	for k, v := range map[string]WrapFn{
+		"GtkFlowBox":      wrapFlowBox,
+		"GtkFlowBoxChild": wrapFlowBoxChild,
+		"GtkPopover":      wrapPopover,
+	} {
+		WrapMap[k] = v
+	}
 }
 
 // GetLocaleDirection() is a wrapper around gtk_get_locale_direction().
@@ -352,4 +375,46 @@ func (fbc *FlowBoxChild) IsSelected() bool {
 // Changed is a wrapper around gtk_flow_box_child_changed()
 func (fbc *FlowBoxChild) Changed() {
 	C.gtk_flow_box_child_changed(fbc.native())
+}
+
+/*
+* GtkPopover
+*/
+
+// Popover is a representation of GTK's GtkPopover.
+type Popover struct {
+	Bin
+}
+
+func (v *Popover) native() *C.GtkPopover {
+	if v == nil || v.GObject == nil {
+		return nil
+	}
+
+	p := unsafe.Pointer(v.GObject)
+	return C.toGtkPopover(p)
+}
+
+func marshalPopover(p uintptr) (interface{}, error) {
+	c := C.g_value_get_object((*C.GValue)(unsafe.Pointer(p)))
+	return wrapPopover(glib.Take(unsafe.Pointer(c))), nil
+}
+
+func wrapPopover(obj *glib.Object) *Popover {
+	return &Popover{Bin{Container{Widget{glib.InitiallyUnowned{obj}}}}}
+}
+
+// PopoverNew is a wrapper around gtk_popover_new().
+func PopoverNew(relative IWidget) (*Popover, error) {
+	//Takes relative to widget
+	var c *C.struct__GtkWidget
+	if relative == nil {
+		c = C.gtk_popover_new(nil)
+	} else {
+		c = C.gtk_popover_new(relative.toWidget())
+	}
+	if c == nil {
+		return nil, nilPtrErr
+	}
+	return wrapPopover(glib.Take(unsafe.Pointer(c))), nil
 }
